+++ src/main/java/roomescape/reservation/controller/ReservationController.java
@@ -10,8 +10,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.reservation.controller.request.CreateReservationRequest;
-import roomescape.reservation.controller.response.FindReservationResponse;
+import roomescape.reservation.dto.request.CreateReservationRequest;
+import roomescape.reservation.dto.response.FindReservationResponse;
 import roomescape.reservation.service.ReservationService;
 
 @RestController
