+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,8 +3,8 @@ package roomescape.reservation.service;
 import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
-import roomescape.reservation.controller.request.CreateReservationRequest;
-import roomescape.reservation.controller.response.FindReservationResponse;
+import roomescape.reservation.dto.request.CreateReservationRequest;
+import roomescape.reservation.dto.response.FindReservationResponse;
 import roomescape.reservation.model.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.model.ReservationTime;
