+++ src/test/java/roomescape/reservation/repository/JdbcReservationRepositoryTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import roomescape.util.DummyDataFixture;
 import roomescape.reservation.model.Reservation;
+import roomescape.util.DummyDataFixture;
 
 @SpringBootTest
 class JdbcReservationRepositoryTest extends DummyDataFixture {
