+++ src/test/java/roomescape/reservation/service/ReservationServiceTest.java
@@ -15,9 +15,9 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
-import roomescape.reservation.controller.request.CreateReservationRequest;
-import roomescape.reservation.controller.response.FindReservationResponse;
-import roomescape.reservation.controller.response.FindTimeOfReservationsResponse;
+import roomescape.reservation.dto.request.CreateReservationRequest;
+import roomescape.reservation.dto.response.FindReservationResponse;
+import roomescape.reservation.dto.response.FindTimeOfReservationsResponse;
 import roomescape.reservation.model.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.reservationtime.model.ReservationTime;
