+++ src/test/java/roomescape/reservationtime/repository/JdbcReservationTimeRepositoryTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import roomescape.util.DummyDataFixture;
 import roomescape.reservationtime.model.ReservationTime;
+import roomescape.util.DummyDataFixture;
 
 @SpringBootTest
 class JdbcReservationTimeRepositoryTest extends DummyDataFixture {
