+++ src/test/java/roomescape/reservationtime/service/ReservationTimeServiceTest.java
@@ -13,8 +13,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
-import roomescape.reservationtime.controller.request.CreateReservationTimeRequest;
-import roomescape.reservationtime.controller.response.FindReservationTimeResponse;
+import roomescape.reservationtime.dto.request.CreateReservationTimeRequest;
+import roomescape.reservationtime.dto.response.FindReservationTimeResponse;
 import roomescape.reservationtime.model.ReservationTime;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
 import roomescape.util.DummyDataFixture;
